Add fixed size queue to track latest processed swap primitives
