+++ protocols/swap/swap.go
@@ -6,13 +6,9 @@ import (
 	"math/big"
 	"strings"
 
-	ethAbi "github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/statechannels/go-nitro/abi"
 	"github.com/statechannels/go-nitro/channel"
 	"github.com/statechannels/go-nitro/channel/state"
-	nc "github.com/statechannels/go-nitro/crypto"
 	"github.com/statechannels/go-nitro/protocols"
 	"github.com/statechannels/go-nitro/types"
 )
@@ -30,126 +26,16 @@ const (
 
 const ObjectivePrefix = "Swap-"
 
-type Exchange struct {
-	TokenIn   common.Address
-	TokenOut  common.Address
-	AmountIn  *big.Int
-	AmountOut *big.Int
-}
-
-func (ex Exchange) Equal(target Exchange) bool {
-	return ex.TokenIn == target.TokenIn && ex.TokenOut == target.TokenOut && ex.AmountIn.Cmp(target.AmountIn) == 0 && ex.AmountOut.Cmp(target.AmountOut) == 0
-}
-
 type SwapPayload struct {
-	SwapPrimitive SwapPrimitive
+	SwapPrimitive channel.SwapPrimitive
 	StateSigs     map[uint]state.Signature
-}
-
-type SwapPrimitive struct {
-	ChannelId types.Destination
-	Exchange  Exchange
-	Sigs      map[uint]state.Signature // keyed by participant index in swap channel
-	Nonce     uint64
-}
-
-func NewSwapPrimitive(channelId types.Destination, tokenIn, tokenOut common.Address, amountIn, amountOut *big.Int, nonce uint64) SwapPrimitive {
-	return SwapPrimitive{
-		ChannelId: channelId,
-		Exchange: Exchange{
-			tokenIn,
-			tokenOut,
-			amountIn,
-			amountOut,
-		},
-		Sigs:  make(map[uint]state.Signature, 2),
-		Nonce: nonce,
-	}
-}
-
-// TODO: Check need of custom marshall and unmarshall methods for swap primitive
-
-// encodes the state into a []bytes value
-func (sp SwapPrimitive) encode() (types.Bytes, error) {
-	// TODO: Check whether we need to encode array of swap primitive
-	// TODO: Check need of app data for sad path will be array of swap primitive
-	return ethAbi.Arguments{
-		{Type: abi.Destination}, // channel id
-		{Type: abi.Address},     // tokenIn
-		{Type: abi.Address},     // tokenOut
-		{Type: abi.Uint256},     // amountIn
-		{Type: abi.Uint256},     // amountOut
-		{Type: abi.Uint256},     // nonce
-	}.Pack(
-		sp.ChannelId,
-		sp.Exchange.TokenIn,
-		sp.Exchange.TokenOut,
-		sp.Exchange.AmountIn,
-		sp.Exchange.AmountOut,
-		new(big.Int).SetUint64(sp.Nonce),
-	)
-}
-
-func (sp SwapPrimitive) Equal(target SwapPrimitive) bool {
-	return sp.ChannelId == target.ChannelId && sp.Exchange.Equal(target.Exchange) && sp.Nonce == target.Nonce
-}
-
-func (sp SwapPrimitive) Clone() SwapPrimitive {
-	clonedSigs := make(map[uint]state.Signature, len(sp.Sigs))
-	for i, sig := range sp.Sigs {
-		clonedSigs[i] = sig
-	}
-
-	return SwapPrimitive{
-		ChannelId: sp.ChannelId,
-		Exchange: Exchange{
-			TokenIn:   sp.Exchange.TokenIn,
-			TokenOut:  sp.Exchange.TokenOut,
-			AmountIn:  sp.Exchange.AmountIn,
-			AmountOut: sp.Exchange.AmountOut,
-		},
-		Sigs:  clonedSigs,
-		Nonce: sp.Nonce,
-	}
-}
-
-// Hash returns the keccak256 hash of the State
-func (sp SwapPrimitive) Hash() (types.Bytes32, error) {
-	encoded, err := sp.encode()
-	if err != nil {
-		return types.Bytes32{}, fmt.Errorf("failed to encode swap primitive: %w", err)
-	}
-	return crypto.Keccak256Hash(encoded), nil
-}
-
-// Sign generates an ECDSA signature on the swap primitive using the supplied private key
-func (sp SwapPrimitive) Sign(secretKey []byte) (state.Signature, error) {
-	hash, error := sp.Hash()
-	if error != nil {
-		return state.Signature{}, error
-	}
-	return nc.SignEthereumMessage(hash.Bytes(), secretKey)
-}
-
-func (sp SwapPrimitive) AddSignature(sig state.Signature, myIndex uint) error {
-	sp.Sigs[myIndex] = sig
-	return nil
-}
-
-// RecoverSigner computes the Ethereum address which generated Signature sig on State state
-func (sp SwapPrimitive) RecoverSigner(sig state.Signature) (types.Address, error) {
-	hash, error := sp.Hash()
-	if error != nil {
-		return types.Address{}, error
-	}
-	return nc.RecoverEthereumMessageSigner(hash[:], sig)
 }
 
 // Objective is a cache of data computed by reading from the store. It stores (potentially) infinite data.
 type Objective struct {
 	Status        protocols.ObjectiveStatus
 	C             *channel.SwapChannel
-	SwapPrimitive SwapPrimitive
+	SwapPrimitive channel.SwapPrimitive
 	StateSigs     map[uint]state.Signature
 	SwapperIndex  uint
 }
@@ -178,7 +64,7 @@ func NewObjective(request ObjectiveRequest, preApprove bool, isSwapper bool, get
 	}
 
 	if isSwapper {
-		swapPrimitive := NewSwapPrimitive(request.channelId, request.tokenIn, request.tokenOut, request.amountIn, request.amountOut, request.nonce)
+		swapPrimitive := channel.NewSwapPrimitive(request.channelId, request.tokenIn, request.tokenOut, request.amountIn, request.amountOut, request.nonce)
 		obj.SwapPrimitive = swapPrimitive
 		obj.SwapperIndex = uint(myIndex)
 	} else {
@@ -328,6 +214,8 @@ func (o *Objective) Crank(secretKey *[]byte) (protocols.Objective, protocols.Sid
 		return &updated, protocols.SideEffects{}, WaitingForSwapping, fmt.Errorf("error updating swap channel state %w", err)
 	}
 
+	// TODO: Add swap primitives to swap channel
+
 	// Completion
 	updated.Status = protocols.Completed
 	return &updated, sideEffects, WaitingForNothing, nil
@@ -440,7 +328,7 @@ func (o *Objective) FindParticipantIndex(address common.Address) (int, error) {
 type jsonObjective struct {
 	Status        protocols.ObjectiveStatus
 	C             types.Destination
-	SwapPrimitive SwapPrimitive
+	SwapPrimitive channel.SwapPrimitive
 	SwapperIndex  uint
 	Nonce         uint64
 	StateSigs     map[uint]state.Signature
@@ -546,6 +434,7 @@ func (r ObjectiveRequest) Id(myAddress types.Address, chainId *big.Int) protocol
 	return protocols.ObjectiveId(ObjectivePrefix + getSwapPrimitiveId(r.fixedPart, r.nonce).String())
 }
 
+// TODO: Discuss swap objective id created from hash of swap primitive
 func getSwapPrimitiveId(fixedPart state.FixedPart, nonce uint64) types.Destination {
 	fp := state.FixedPart{
 		Participants:      fixedPart.Participants,
